Move SimpleDOM test to Environment tests
